1003・1004の修正

+++ Models/GW1004/Response/Tesuryo.cs
@@ -2,11 +2,11 @@
 {
     public class Tesuryo
     {
-        public int sofuriToko3manemMimanTesuryo { get; set; }
-        public int sofuriToko3manenIjyoTesuryo { get; set; }
-        public int sofuriTako3manemMimanTesuryo { get; set; }
-        public int sofuriTako3manemIjyoTesuryo { get; set; }
-        public int kyufuriTokoTesuryo { get; set; }
-        public int kyufuriTakoTesuryo { get; set; }
+        public long sofuriToko3manemMimanTesuryo { get; set; }
+        public long sofuriToko3manenIjyoTesuryo { get; set; }
+        public long sofuriTako3manemMimanTesuryo { get; set; }
+        public long sofuriTako3manemIjyoTesuryo { get; set; }
+        public long kyufuriTokoTesuryo { get; set; }
+        public long kyufuriTakoTesuryo { get; set; }
     }
 }
